projet/enonce/requirements.txt
confluent-kafka
cassandra-driver
requests
streamlit
pandas
plotly
folium
streamlit-folium
